+++ ChangeLog.md
@@ -1,5 +1,12 @@
 2020-05
 
+- adding NBayes learning and classification.
+
+       $ ecoz2 nbayes learn data/sequences/TRAIN/M2048/B
+
+    As part of this, new `sequence` module to load and display sequences
+    generated from C version.
+
 - 0.3.63 - align with C
 - release using github actions (linux and macos binaries)